pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "difffit-3d"
version = "0.1.0"
description = "Geometry-Aware Virtual Try-On: Synergizing 3D Garment Fitting with Controlled Diffusion Models"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.9"
authors = [
    {name = "DiffFit-3D Team"},
]
keywords = ["virtual-try-on", "diffusion", "3d", "garment-fitting", "computer-vision"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "diffusers>=0.25.0",
    "transformers>=4.36.0",
    "accelerate>=0.25.0",
    "safetensors>=0.4.0",
    "xformers>=0.0.23",
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "scikit-image>=0.21.0",
    "einops>=0.7.0",
    "omegaconf>=2.3.0",
    "wandb>=0.16.0",
    "tqdm>=4.66.0",
    "lpips>=0.1.4",
    "pytorch-fid>=0.3.0",
    "kornia>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]
serve = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "celery[redis]>=5.3.0",
    "Pillow>=10.0.0",
]
threed = [
    "pytorch3d>=0.7.5",
    "trimesh>=4.0.0",
    "pyrender>=0.1.45",
    "smplx>=0.1.28",
]

[project.urls]
Homepage = "https://github.com/DiffFit-3D/DiffFit-3D"
Repository = "https://github.com/DiffFit-3D/DiffFit-3D"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --tb=short"
